requirements.txt
fastapi
uvicorn
pytesseract
pillow
rapidfuzz
ollama
sqlalchemy
python-multipart